# server/swagger-spec.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/notion/databases": {
      "get": {
        "operationId": "NotionDatabaseController_getDatabases",
        "summary": "Get all databases for the authenticated user",
        "description": "Returns a list of Notion databases the user has access to. Can be filtered by workspace ID.",
        "parameters": [
          {
            "name": "workspace",
            "required": false,
            "in": "query",
            "description": "Filter databases by workspace ID",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "List of databases returned successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/NotionDatabaseDto"
                  }
                }
              }
            }
          },
          "401": {
            "description": "Not authenticated"
          },
          "403": {
            "description": "Not authorized (subscription required)"
          }
        },
        "tags": [
          "notion-databases"
        ],
        "security": [
          {
            "JWT-auth": []
          }
        ]
      }
    },
    "/notion/databases/{id}": {
      "get": {
        "operationId": "NotionDatabaseController_getDatabase",
        "summary": "Get a specific database by ID",
        "description": "Returns detailed information about a single Notion database.",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "The ID of the database to retrieve",
            "schema": {
              "example": "db_123456789",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Database returned successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/NotionDatabaseDto"
                }
              }
            }
          },
          "401": {
            "description": "Not authenticated"
          },
          "403": {
            "description": "Not authorized (subscription required)"
          },
          "404": {
            "description": "Database not found"
          }
        },
        "tags": [
          "notion-databases"
        ],
        "security": [
          {
            "JWT-auth": []
          }
        ]
      }
    },
    "/notion/databases/{id}/sync": {
      "post": {
        "operationId": "NotionDatabaseController_syncDatabase",
        "summary": "Sync a database with Notion",
        "description": "Updates the local copy of the database with the latest data from Notion.",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "The ID of the database to sync",
            "schema": {
              "example": "db_123456789",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Database sync initiated successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DatabaseSyncResultDto"
                }
              }
            }
          },
          "401": {
            "description": "Not authenticated"
          },
          "403": {
            "description": "Not authorized (subscription required)"
          },
          "404": {
            "description": "Database not found"
          }
        },
        "tags": [
          "notion-databases"
        ],
        "security": [
          {
            "JWT-auth": []
          }
        ]
      }
    },
...
  },
  "info": {
    "title": "Notion Graph View API",
    "description": "API Documentation for the Notion Graph View application",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "JWT-auth": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http",
        "description": "Enter JWT token"
      }
    },
    "schemas": {
...
      "NotionDatabaseDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "description": "Unique identifier for the database",
            "example": "db_123456789"
          },
          "title": {
            "type": "string",
            "description": "Title of the database",
            "example": "Project Tasks"
          },
          "workspaceId": {
            "type": "string",
            "description": "ID of the workspace that contains the database",
            "example": "ws_123456789"
          },
          "ownerId": {
            "type": "string",
            "description": "ID of the database owner",
            "example": "user_123456789"
          },
          "lastSyncedAt": {
            "format": "date-time",
            "type": "string",
            "description": "Timestamp of the last synchronization with Notion",
            "example": "2023-01-15T12:30:45.123Z"
          },
          "description": {
            "type": "string",
            "description": "Optional description of the database",
            "example": "This database contains all project tasks"
          },
          "url": {
            "type": "string",
            "description": "URL to the database in Notion",
            "example": "https://notion.so/workspace/db-123456789"
          },
          "createdAt": {
            "format": "date-time",
            "type": "string",
            "description": "Timestamp when the database was created",
            "example": "2023-01-01T10:00:00.000Z"
          },
          "updatedAt": {
            "format": "date-time",
            "type": "string",
            "description": "Timestamp when the database was last updated",
            "example": "2023-01-15T12:30:45.123Z"
          },
          "pageCount": {
            "type": "number",
            "description": "Number of pages in the database",
            "example": 42
          }
        },
        "required": [
          "id",
          "title",
          "workspaceId",
          "ownerId",
          "lastSyncedAt",
          "createdAt",
          "updatedAt"
        ]
      },
      "DatabaseSyncResultDto": {
        "type": "object",
        "properties": {
          "databaseId": {
            "type": "string",
            "description": "ID of the database that was synced",
            "example": "db_123456789"
          },
          "newPages": {
            "type": "number",
            "description": "Number of new pages created during sync",
            "example": 5
          },
          "updatedPages": {
            "type": "number",
            "description": "Number of existing pages updated during sync",
            "example": 10
          },
          "totalPages": {
            "type": "number",
            "description": "Total number of pages in the database",
            "example": 42
          },
          "syncedAt": {
            "format": "date-time",
            "type": "string",
            "description": "Timestamp when the sync was completed",
            "example": "2023-01-15T12:30:45.123Z"
          }
        },
        "required": [
          "databaseId",
          "newPages",
          "updatedPages",
          "totalPages",
          "syncedAt"
        ]
      },
...
    }
  }
}